Move decodable conformance for exposure window to test target
